+++ philo/utils.c
@@ -6,16 +6,16 @@
 /*   By: molasz-a <molasz-a@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/29 11:25:40 by molasz-a          #+#    #+#             */
-/*   Updated: 2024/06/02 17:37:59 by molasz-a         ###   ########.fr       */
+/*   Updated: 2024/06/02 17:59:59 by molasz-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-void	ft_sleep(long mili)
+void	ft_sleep(size_t mili)
 {
-	long	start;
-	long	time;
+	size_t	start;
+	size_t	time;
 
 	start = get_time();
 	while (1)
@@ -26,7 +26,7 @@ void	ft_sleep(long mili)
 	}
 }
 
-long	get_time(void)
+size_t	get_time(void)
 {
 	struct timeval	time;
 
@@ -34,7 +34,7 @@ long	get_time(void)
 	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
 
-void	print_time_philo(long time, int philo)
+void	print_time_philo(size_t time, int philo)
 {
 	putnbr_fd(time, 1);
 	putchar_fd(' ', 1);
